Build IPCA for Linux.
Reviewed-on: https://gerrit.iotivity.org/gerrit/18409
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-on: https://gerrit.iotivity.org/gerrit/19079

+++ resource/IPCA/src/ipcavariant.cpp
@@ -131,7 +131,7 @@ IPCAStatus IPCAPropertyBagSetValueArray(IPCAPropertyBagHandle propertyBagHandle,
     OC::OCRepresentation* ocRep = reinterpret_cast<OC::OCRepresentation*>(propertyBagHandle);
 
     std::vector<_T> array;
-    for (int i = 0; i < arrayCount; i++)
+    for (size_t i = 0; i < arrayCount; i++)
     {
         array.insert(array.end(), valueArray[i]);
     }
@@ -187,7 +187,7 @@ IPCAStatus IPCAPropertyBagSetValueStringArray(IPCAPropertyBagHandle propertyBagH
     OC::OCRepresentation* ocRep = reinterpret_cast<OC::OCRepresentation*>(propertyBagHandle);
 
     std::vector<std::string> array;
-    for (int i = 0; i < arrayCount; i++)
+    for (size_t i = 0; i < arrayCount; i++)
     {
         array.insert(array.end(), valueArray[i]);
     }
@@ -218,7 +218,7 @@ IPCAStatus IPCAPropertyBagSetValuePropertyBagArray(IPCAPropertyBagHandle propert
     OC::OCRepresentation* ocRep = reinterpret_cast<OC::OCRepresentation*>(propertyBagHandle);
 
     std::vector<OC::OCRepresentation> ocRepArray;
-    for (int i = 0; i < valueCount; i++)
+    for (size_t i = 0; i < valueCount; i++)
     {
         ocRepArray.insert(ocRepArray.end(),
             *(reinterpret_cast<OC::OCRepresentation*>((valueArray[i]))));
@@ -387,7 +387,7 @@ IPCAStatus IPCAPropertyBagGetValuePropertyBagArray(IPCAPropertyBagHandle propert
         return IPCA_OUT_OF_MEMORY;
     }
 
-    int i = 0;
+    size_t i = 0;
     IPCAStatus status = IPCA_FAIL;
     for (auto& ocrep : array)
     {
@@ -421,7 +421,7 @@ IPCAStatus IPCAPropertyBagGetValuePropertyBagArray(IPCAPropertyBagHandle propert
 
 void IPCAPropertyBagFreePropertyBagArray(IPCAPropertyBagHandle* valueArray, size_t valueCount)
 {
-    for (int i = 0; i < valueCount; i++)
+    for (size_t i = 0; i < valueCount; i++)
     {
         IPCAPropertyBagDestroy(valueArray[i]);
     }
